\name{normwn.test-package}
\alias{normwn.test-package}
\alias{normwn.test}
\docType{package}
\title{ Normality and White Noise Testing
}
\description{
Performs omnibus univariate and multivariate tests for normality as developed by Doornik and Hansen (1994) designed to deal with the small sample problems associated with the Jarque-Bera tests. A variant of the tests deal with the problems posed if variables are not independent but weakly dependent. Also included is an univariate test for white noise which similarly allows for weak dependence.
}
\details{
\tabular{ll}{
Package: \tab normwn.test\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-02-07\cr
License: \tab GPL version 3\cr
}

}
\author{
Peter Wickham <peterwickham@mac.com>

Maintainer: 
Peter Wickham <peterwickham@mac.com>
}
\references{
Doornik, J.A., and H. Hansen (1994). "An Omnibus Test for Univariate and Multivariate Normality". Working Paper, Nuffield College, Oxford University, Oxford, U.K.
Lobato, I., and C. Velasco (2004). "A Simple Test of Normality of Time Series". Econometric Theory, 20, 671-689, Cambridge University Press.
Lobato, I., and C. Velasco (2004). "A Simple and General Test for White Noise". Econometric Society, Latin-America Meetings, Paper No. 112
}
\keyword{ package }
\seealso{

}
\examples{

}
